Update jdbc dao and junit test controllers

+++ src/main/java/com/example/catalog/controllers/BookController.java
@@ -1,13 +1,17 @@
 package com.example.catalog.controllers;
 
+import com.example.catalog.exception.EntityNotFoundException;
 import com.example.catalog.model.Book;
 import com.example.catalog.service.BookService;
+import com.example.catalog.util.ModelUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.ArrayList;
+import java.util.List;
 
 
 @RestController
@@ -18,22 +22,31 @@ public class BookController {
     private BookService bookService;
 
     @GetMapping(path = "/getBook")
-    public ArrayList<Book> getBook() {
-        return (ArrayList<Book>) bookService.findAll();
+    public ResponseEntity<List<Book>> getBooks() {
+        List<Book> bookList =  bookService.findAll();
+        return ResponseEntity.ok().body(bookList);
+    }
+
+    @GetMapping(value = "/{id}")
+    public ResponseEntity<Book> getOneBook (@PathVariable("id") Long id) throws EntityNotFoundException{
+        if (ModelUtils.idIsNotPresent(id))
+            throw new EntityNotFoundException("id-" + id);
+        Book book = bookService.findOne(id);
+        return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(book);
     }
 
     @PostMapping(path = "/addBook")
     @ResponseBody
-    public ArrayList<Book> addBook(@Valid @RequestBody Book requestBook) {
+    public ResponseEntity<Book> addBook(@Valid @RequestBody Book requestBook) {
         Book book = new Book(requestBook.getIsbn(), requestBook.getName(), requestBook.getAuthor(),
                 requestBook.getNumberOfPages(), requestBook.getWeight(), requestBook.getPrice());
         bookService.save(book);
-        return (ArrayList<Book>) bookService.findAll();
+        return ResponseEntity.status(201).body(book);
     }
 
     @PutMapping(path = "/editBook/{id}")
     @ResponseBody
-    public ArrayList<Book> editBook(@PathVariable(value = "id") long id,@Valid @RequestBody Book requestBook) {
+    public ResponseEntity<Book> editBook(@PathVariable(value = "id") long id,@Valid @RequestBody Book requestBook) throws EntityNotFoundException {
         Book book = bookService.findOne(id);
         book.setIsbn(requestBook.getIsbn());
         book.setName(requestBook.getName());
@@ -42,14 +55,16 @@ public class BookController {
         book.setWeight(requestBook.getWeight());
         book.setPrice(requestBook.getPrice());
         bookService.save(book);
-        return (ArrayList<Book>) bookService.findAll();
+        return ResponseEntity.ok().body(book);
     }
 
     @DeleteMapping(path = "/deleteBook/{id}")
     @ResponseBody
-    public ArrayList<Book> deleteBookBook(@PathVariable(value = "id") long id) {
+    public ResponseEntity<List<Book>> deleteBookBook(@PathVariable(value = "id") long id) throws EntityNotFoundException{
+        if (ModelUtils.idIsNotPresent(id))
+            throw new EntityNotFoundException("id-" + id);
         bookService.remove(id);
-        return (ArrayList<Book>) bookService.findAll();
+        return ResponseEntity.ok().body(bookService.findAll());
     }
 
 }
